[refactor] move class and methods to jnlp package
