+++ exist-core/src/main/java/org/exist/webstart/JnlpJarFiles.java
@@ -33,7 +33,6 @@ import java.util.stream.Collectors;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.exist.start.LatestFileResolver;
 import org.exist.util.FileUtils;
 
 /**
